refactor: change scraper name from playwright_amazon to amazon
