+++ amazon/amazon/settings.py
@@ -31,9 +31,9 @@ POSTGRES_HOST = getenv("POSTGRES_HOST") or "127.0.0.1"
 # end .env variables
 
 
-BOT_NAME = "playwright_amazon"
-SPIDER_MODULES = ["playwright_amazon.spiders"]
-NEWSPIDER_MODULE = "playwright_amazon.spiders"
+BOT_NAME = "amazon"
+SPIDER_MODULES = ["amazon.spiders"]
+NEWSPIDER_MODULE = "amazon.spiders"
 
 
 # Obey robots.txt rules
@@ -45,14 +45,14 @@ CONCURRENT_REQUESTS = 16
 # Enable or disable downloader middlewares
 # See https://docs.scrapy.org/en/latest/topics/downloader-middleware.html
 DOWNLOADER_MIDDLEWARES = {
-    "playwright_amazon.middlewares.FakeHeaderMiddleware": 400,
+    "amazon.middlewares.FakeHeaderMiddleware": 400,
 }
 
 # Configure item pipelines
 # See https://docs.scrapy.org/en/latest/topics/item-pipeline.html
 ITEM_PIPELINES = {
-    "playwright_amazon.pipelines.PlaywrightAmazonPipeline": 300,
-    "playwright_amazon.pipelines.SaveToPostgresPipeline": 400,
+    "amazon.pipelines.PlaywrightAmazonPipeline": 300,
+    "amazon.pipelines.SaveToPostgresPipeline": 400,
 }
 
 # Set settings whose default value is deprecated to a future-proof value
