+++ amazon/scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = playwright_amazon.settings
+default = amazon.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = playwright_amazon
+project = amazon
